Create logic.js
To distribute the drug, the first distributors enters the “Ready” state and the second distributor enter the “Drug Demand” state. Then, when GPS and temperature sensors receive the drugs, the transaction is automatically generate. If the drug condition (temperature, GPS and the paid) is not in the defined value, the distributor is penalised. Additonally, TTL, Drugtime, and DrugLoc values are updated.